Add GameController.py module
Refactored general game logic by moving the RainingCoins class to a new module, GameController.py, in the game_state folder.
